Extract constant literal emission from TransformWorkerProgram
Move the listed members out of TransformWorkerProgram without changing their bodies.
